Change logo to img and move svg icon components to components dir
