+++ src/styles/base.css
@@ -10,7 +10,7 @@
   margin-inline-start: 0;
   margin-inline-end: 0;
 
-  --color-primary: rgb(146, 35, 44);
+  --color-primary: #92232C;
   --color-magnolia: #f7f7f7;
   --color-disabled: dimgrey;
 }
